finish norm check, write Makefile

+++ ft_strlcpy.c
@@ -1,17 +1,29 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strlcpy.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: yenoh <yenoh@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/05/09 08:04:05 by yenoh             #+#    #+#             */
+/*   Updated: 2023/05/09 08:38:07 by yenoh            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
-int ft_strlcpy(char *dst, const char *src, size_t dstsize)
+int	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
-    int i;
-    int len;
+	size_t	i;
+	int		len;
 
-    i = 0;
-    len = ft_strlen(src);
-    while (src[i] && i < dstsize - 1)
-    {
-        dst[i] = src[i];
-        i++;
-    }
-    dst[i] = 0;
-    return (len);
+	i = 0;
+	len = ft_strlen(src);
+	while (src[i] && i < dstsize - 1)
+	{
+		dst[i] = src[i];
+		i++;
+	}
+	dst[i] = 0;
+	return (len);
 }
